requirements.txt
fabric>=3.0
anthropic>=0.40
requests>=2.31
pyyaml>=6.0
pandas>=2.0
matplotlib>=3.7
pytest>=7.0
